requirements.txt
Flask
Pillow
fpdf
python-docx
pandas
openpyxl
PyMuPDF
pdf2image
pdf2docx
docx2pdf
tabula-py